CR fixes - docs and rename to policy_*

+++ clients/drcachesim/simulator/cache.h
@@ -63,8 +63,8 @@ public:
     // The id is an index into the snoop filter's array of caches for coherent caches.
     // If this is a coherent cache, id should be in the range [0,num_snooped_caches).
     bool
-    init(int associativity, int64_t line_size, int total_size, caching_device_t *parent,
-         caching_device_stats_t *stats,
+    init(int associativity, int64_t line_size, int64_t total_size,
+         caching_device_t *parent, caching_device_stats_t *stats,
          std::unique_ptr<cache_replacement_policy_t> replacement_policy,
          prefetcher_t *prefetcher = nullptr,
          cache_inclusion_policy_t inclusion_policy =
